+++ packages/storefront/pages/order/index.tsx
@@ -1,10 +1,8 @@
 import { GetServerSideProps, NextPage } from 'next';
 var cookie = require('cookie');
-import Breadcrumb from '@/components/global/breadcrumbs/breadcrumb';
 import { userAPI } from 'APIs';
 import { IOrderResponseData } from 'models';
-import DataTable from '@/components/order/dataTable';
-import Link from 'next/link';
+import OrderMain from '@/components/order/ordersList/index';
 
 interface Props {
   orderProducts: {
@@ -19,33 +17,7 @@ const Order: NextPage<Props> = ({ orderProducts }: Props) => {
 
   return (
     <>
-      <Breadcrumb
-        title="My order"
-        pathArray={['Home', 'My order']}
-        linkArray={['/', '/']}
-      />
-      <section className="container mx-auto px-4 ">
-        <div className="flex flex-col items-center border-b py-6">
-          <div
-            className={
-              storedOrderProducts?.length
-                ? 'mb-2 w-full overflow-x-scroll xl:overflow-x-hidden'
-                : 'mb-10'
-            }
-          >
-            {storedOrderProducts?.length ? (
-              <DataTable storedOrderProducts={storedOrderProducts} />
-            ) : (
-              'You have not placed any order yet'
-            )}
-          </div>
-          <div className="mt-1" style={{ textAlign: 'center' }}>
-            <button className="rounded-md bg-green-600 py-4 px-6 font-light text-white transition-all duration-200 ease-linear hover:bg-stone-900">
-              <Link href="/">Go to home page</Link>
-            </button>
-          </div>
-        </div>
-      </section>
+      <OrderMain orderProducts={orderProducts} />
     </>
   );
 };
